Revert "treewide(feat): account for filesystem paths"
This method of checking for CAF-given names is too unreliable and hacky.
A better way to do this would be to modify the manifest itself to
include a custom `caf` attribute in `project` tags. Let's do that
instead. This reverts commit eeafc344632ad66fe88f048cd4b6ad0aba7e3137.